Clean structure and rename files
