+++ src/helper/spotify.ts
@@ -1,2 +1,2 @@
 export const getTokenFromUrl = (): string | undefined =>
-  window.location.hash.split('=')[1].split('&')[0]
+  window.location.hash.split('=')[1]?.split('&')[0]
